+++ src/lib.rs
@@ -74,7 +74,7 @@ pub use arbiter::Arbiter;
 pub use address::{Address, SyncAddress, Subscriber, ActorAddress};
 pub use context::Context;
 pub use envelope::ToEnvelope;
-pub use framed::FramedContext;
+pub use framed::{FramedCell, FramedContext};
 pub use message::{Request, Response};
 pub use sync::{SyncContext, SyncArbiter};
 pub use registry::{Registry, SystemRegistry, ArbiterService, SystemService};